requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-google-genai==0.0.11
google-generativeai==0.3.2
python-dotenv==1.0.1 